backend/requirements.txt
alembic
fastapi
passlib[bcrypt]
psycopg2-binary
pydantic[email]
python-dotenv
python-jose[cryptography]
python-multipart
sqlalchemy
uvicorn[standard]
pandas
openpyxl